requirements.txt
# 基础依赖
numpy>=1.19.0
torch>=1.9.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.7b0
flake8>=3.9.2
isort>=5.9.3
mypy>=0.910
opencv-python>=4.5.0
face-recognition>=1.3.0
dlib>=19.22.0
PyYAML>=5.4.0
pillow>=8.3.1
tqdm>=4.62.0
python-dotenv>=0.19.0

# 语音处理
librosa>=0.9.0
sounddevice>=0.4.4
scipy>=1.7.0

# 数据库
SQLAlchemy>=1.4.0
alembic>=1.7.0

# 机器学习
scikit-learn>=0.24.0
torchvision>=0.10.0

# 语音合成（可选）
pyttsx3>=2.90

# 工具
tqdm>=4.62.0
python-dotenv>=0.19.0 